Break es6 code out into files
